+++ src/components/molecules/Selector/index.jsx
@@ -3,6 +3,7 @@ import Icon from 'components/atoms/Icon';
 import * as Styled from 'components/molecules/Selector/Selector.style';
 import mockData from 'components/molecules/Selector//SelectMockData';
 import ListItem from 'components/atoms/ListItem';
+import Button from 'components/atoms/Button';
 
 const Selector = ({ initTitle = 'Select', listData = mockData, ...props }) => {
   const [isOpen, setIsOpen] = useState(false);
@@ -20,22 +21,31 @@ const Selector = ({ initTitle = 'Select', listData = mockData, ...props }) => {
     fontType: 'large',
   };
 
+  const buttonStyle = {
+    sizeType: 'small',
+    colorType: 'point',
+    fontType: 'medium',
+  };
+
   return (
-    <Styled.Selector>
-      <Styled.SelectBox flexType="centerBetween" borderType="default" onClick={toggleIsOpen}>
-        <Icon iconName={isOpen ? 'caretUp' : 'caretDown'} />
-        <Styled.SelectTitle>{selectedValue.title}</Styled.SelectTitle>
-      </Styled.SelectBox>
-      {isOpen && (
-        <Styled.SelectList>
-          {listData.map(selectedData => (
-            <ListItem key={selectedData.id} onClick={clickItem(selectedData)} {...ListItemStyle}>
-              {selectedData.title}
-            </ListItem>
-          ))}
-        </Styled.SelectList>
-      )}
-    </Styled.Selector>
+    <Styled.SelectorWrapper>
+      <Styled.Selector>
+        <Styled.SelectBox flexType="centerBetween" borderType="default" onClick={toggleIsOpen}>
+          <Icon iconName={isOpen ? 'caretUp' : 'caretDown'} />
+          <Styled.SelectTitle>{selectedValue.title}</Styled.SelectTitle>
+        </Styled.SelectBox>
+        {isOpen && (
+          <Styled.SelectList>
+            {listData.map(selectedData => (
+              <ListItem key={selectedData.id} onClick={clickItem(selectedData)} {...ListItemStyle}>
+                {selectedData.title}
+              </ListItem>
+            ))}
+          </Styled.SelectList>
+        )}
+      </Styled.Selector>
+      <Button {...buttonStyle}>추가</Button>
+    </Styled.SelectorWrapper>
   );
 };
