+++ src/components/molecules/UserInputBox/index.jsx
@@ -1,6 +1,5 @@
 import React, { useContext } from 'react';
 import Input from 'components/atoms/Input';
-import Button from 'components/atoms/Button';
 import Selector from 'components/molecules/Selector';
 import * as Styled from 'components/molecules/UserInputBox/UserInputBox.style';
 import { CURRENCY_STR } from 'constants';
@@ -11,12 +10,6 @@ import { calcInsertedMoney } from 'context/Wallet/reducer';
 const UserInputBox = ({ title, ...props }) => {
   const { state, walletDispatch, insertMoney } = useContext(WalletContext);
   const { addLog, LOG_TYPE, logDispatch } = useContext(LogContext);
-
-  const buttonStyle = {
-    sizeType: 'small',
-    colorType: 'point',
-    fontType: 'medium',
-  };
 
   const handleKeyPress = ({ key, target }) => {
     if (key !== 'Enter' || target.value === '') {
@@ -48,10 +41,7 @@ const UserInputBox = ({ title, ...props }) => {
   return (
     <Styled.UserInputBox>
       <Input onKeyPress={handleKeyPress} onInput={handleOnInput} type="number" />
-      <Styled.SelectorWrapper>
-        <Selector></Selector>
-        <Button {...buttonStyle}>추가</Button>
-      </Styled.SelectorWrapper>
+      <Selector />
     </Styled.UserInputBox>
   );
 };
